gallery/requirements.txt
django==5.0.7
djangorestframework==3.15.2
psycopg2==2.9.9
pillow==10.4.0
django-cloudinary-storage==0.3.0